Update README.md
Violence is a major social problem that can have devastating 
consequences for individuals and communities. Real-time violence detection systems can help to prevent violence by identifying and responding to violent incidents in real time. The proposed real-time violence detection system has the potential to be a valuable tool for preventing violence in a variety of settings, such as schools, workplaces, and public spaces. The objective of this project is to develop a real-time violence detection system that can accurately identify violent incidents in surveillance video streams. It will use a combination of Deep Learning algorithms such as CNN, LSTM and Computer Vision techniques to extract features from surveillance video streams and classify them as violent or non-violent. The system will be trained on a large dataset of labelled surveillance videos. Real 
Life Violent Videos Dataset is used for this purpose which has 1000 violent videos and 1000 non-violent videos. Once trained, the system will be able to detect violence in real time with high accuracy.